Update face match logic
